More work on terrain. Getting there.

+++ src/swganh/tre/visitors/terrain/detail/boundary_layer.h
@@ -14,12 +14,12 @@ namespace tre
 		
 		virtual LayerType GetType() { return LAYER_TYPE_BOUNDARY; }
 
-		virtual bool IsContained(float px, float pz) = 0;
-		virtual float Process(float px, float pz) = 0;
+		virtual bool IsContained(double px, double pz) = 0;
+		virtual double Process(double px, double pz) = 0;
 		
 	protected:
 		uint32_t feather_type;
-		float feather_amount;
+		double feather_amount;
 	};
 }
 }
